Im dumb

+++ src/main/java/club/hellin/vivillyapi/utils/Utils.java
@@ -71,5 +71,5 @@ public abstract class Utils {
 
     public abstract String getInventoryName(final InventoryView view);
 
-    public abstract CourseEntryBase createCourseEntry(final String uuid, final int time, final int rocketsUsed, final int timesHitGround, final int finishedAt);
+    public abstract CourseEntryBase createCourseEntry(final String uuid, final int time, final int rocketsUsed, final int timesHitGround, final long finishedAt);
 }
